Code refactor (smaller) (#3065)
