+++ clients/daco-dialect-support/package.json
@@ -106,8 +106,8 @@
     "build": "npm run compile && esbuild src/extension.ts --bundle --sourcemap --target=node16.17.1,chrome108 --outdir=dist --external:vscode --format=cjs",
     "compile": "npm run grammar && tsc -p ./",
     "grammar": "npm run lexer && npm run parser",
-    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/DaCoLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/CopybookLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/VariableLexer.g4",
-    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/DaCoParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/CopybookParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/VariableParser.g4",
+    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/StatementsLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/ProgramLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/VariableLexer.g4",
+    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/StatementsParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/ProgramParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/VariableParser.g4",
     "watch": "tsc -watch -p ./",
     "lint-format": "prettier -c .",
     "lint-quality": "eslint ./src",
